models files init, improved reactivitiy, confirm dialogs, fixed guards and more

+++ frontend/src/app/users/users.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnChanges, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { Observable } from 'rxjs';
+import { ConfirmDialogComponent } from '../confirm-dialog/confirm-dialog.component';
 import { UserService } from '../shared/data-access/user.service';
 
 @Component({
@@ -9,7 +11,7 @@ import { UserService } from '../shared/data-access/user.service';
 })
 export class UsersComponent implements OnInit {
 
-  constructor(private userService : UserService) { }
+  constructor(private userService : UserService, private dialog : MatDialog) { }
 
   users$ : Observable<any> = this.userService.getUsers()
 
@@ -21,12 +23,16 @@ export class UsersComponent implements OnInit {
   }
 
   onSetDefaultAvatar(id : number) {
-    if (!confirm('Ești sigur?')) return
-    this.userService.setDefaultAvatar(id).subscribe()
+    let confirmDialog = this.dialog.open(ConfirmDialogComponent);
+    confirmDialog.afterClosed().subscribe((res) => {
+      if (res) this.userService.setDefaultAvatar(id).subscribe()
+    })
   }
 
   onDelete(id : number) {
-    if (!confirm('Ești sigur?')) return
-    this.userService.deleteUser(id).subscribe(() => this.users$ = this.userService.getUsers())
+    let confirmDialog = this.dialog.open(ConfirmDialogComponent);
+    confirmDialog.afterClosed().subscribe((res) => {
+      if (res) this.userService.deleteUser(id).subscribe(() => this.users$ = this.userService.getUsers())
+    })
   }
 }
